Commit 3
-Networking! So, it’s semi functional. I tested it using my internet
and it barely works at all because of the absurd latency and quite
frankly I have no clue how to fix that and so I’m gonna pawn off the
blame onto my internet. I really hope it is my internet but I’m
entirely unsure as this is the only thing I’ve tested it on. Any case,
I got the networking working to some extent.